Clean up modern plugins to remove dead code
Remove names
