+++ packages/react-dom/src/events/DOMModernPluginEventSystem.js
@@ -33,7 +33,12 @@ import {
   USE_EVENT_SYSTEM,
 } from './EventSystemFlags';
 
-import {HostRoot, HostPortal} from 'react-reconciler/src/ReactWorkTags';
+import {
+  HostRoot,
+  HostPortal,
+  ScopeComponent,
+  HostComponent,
+} from 'react-reconciler/src/ReactWorkTags';
 
 import {
   addTrappedEventListener,
@@ -86,10 +91,12 @@ import {
 import {COMMENT_NODE} from '../shared/HTMLNodeType';
 import {topLevelEventsToDispatchConfig} from './DOMEventProperties';
 import {batchedEventUpdates} from './ReactDOMUpdateBatching';
+import getListener from './getListener';
 
 import {
   enableLegacyFBSupport,
   enableUseEventAPI,
+  enableScopeAPI,
 } from 'shared/ReactFeatureFlags';
 import {
   invokeGuardedCallbackAndCatchFirstError,
@@ -153,7 +160,7 @@ const isArray = Array.isArray;
 const PossiblyWeakMap = ((typeof WeakMap === 'function' ? WeakMap : Map): any);
 
 // $FlowFixMe: Flow cannot handle polymorphic WeakMaps
-export const eventTargetEventListenerStore: WeakMap<
+const eventTargetEventListenerStore: WeakMap<
   EventTarget,
   Map<
     DOMTopLevelEventType,
@@ -162,7 +169,7 @@ export const eventTargetEventListenerStore: WeakMap<
 > = new PossiblyWeakMap();
 
 // $FlowFixMe: Flow cannot handle polymorphic WeakMaps
-export const reactScopeListenerStore: WeakMap<
+const reactScopeListenerStore: WeakMap<
   ReactScopeMethods,
   Map<
     DOMTopLevelEventType,
@@ -670,5 +677,305 @@ export function detachListenerFromReactScope(listener: ReactDOMListener): void {
         listeners.bubbled.delete(listener);
       }
     }
+  }
+}
+
+export function accumulateTwoPhaseListeners(
+  event: ReactSyntheticEvent,
+  accumulateUseEventListeners?: boolean,
+): void {
+  const phasedRegistrationNames = event.dispatchConfig.phasedRegistrationNames;
+  const dispatchListeners = [];
+  const dispatchInstances: Array<Fiber | null> = [];
+  const dispatchCurrentTargets = [];
+
+  const {bubbled, captured} = phasedRegistrationNames;
+  // If we are not handling EventTarget only phase, then we're doing the
+  // usual two phase accumulation using the React fiber tree to pick up
+  // all relevant useEvent and on* prop events.
+  let instance = event._targetInst;
+  let lastHostComponent = null;
+
+  // Accumulate all instances and listeners via the target -> root path.
+  while (instance !== null) {
+    const {stateNode, tag} = instance;
+    // Handle listeners that are on HostComponents (i.e. <div>)
+    if (tag === HostComponent && stateNode !== null) {
+      const currentTarget = stateNode;
+      lastHostComponent = currentTarget;
+      // For useEvent listenrs
+      if (enableUseEventAPI && accumulateUseEventListeners) {
+        // useEvent event listeners
+        const targetType = event.type;
+        const listeners = getListenersFromTarget(currentTarget);
+
+        if (listeners !== null) {
+          const listenersArr = Array.from(listeners);
+          for (let i = 0; i < listenersArr.length; i++) {
+            const listener = listenersArr[i];
+            const {
+              callback,
+              event: {capture, type},
+            } = listener;
+            if (type === targetType) {
+              if (capture === true) {
+                dispatchListeners.unshift(callback);
+                dispatchInstances.unshift(instance);
+                dispatchCurrentTargets.unshift(currentTarget);
+              } else {
+                dispatchListeners.push(callback);
+                dispatchInstances.push(instance);
+                dispatchCurrentTargets.push(currentTarget);
+              }
+            }
+          }
+        }
+      }
+      // Standard React on* listeners, i.e. onClick prop
+      if (captured !== null) {
+        const captureListener = getListener(instance, captured);
+        if (captureListener != null) {
+          // Capture listeners/instances should go at the start, so we
+          // unshift them to the start of the array.
+          dispatchListeners.unshift(captureListener);
+          dispatchInstances.unshift(instance);
+          dispatchCurrentTargets.unshift(currentTarget);
+        }
+      }
+      if (bubbled !== null) {
+        const bubbleListener = getListener(instance, bubbled);
+        if (bubbleListener != null) {
+          // Bubble listeners/instances should go at the end, so we
+          // push them to the end of the array.
+          dispatchListeners.push(bubbleListener);
+          dispatchInstances.push(instance);
+          dispatchCurrentTargets.push(currentTarget);
+        }
+      }
+    }
+    if (
+      enableUseEventAPI &&
+      enableScopeAPI &&
+      accumulateUseEventListeners &&
+      tag === ScopeComponent &&
+      lastHostComponent !== null
+    ) {
+      const reactScope = stateNode.methods;
+      const eventTypeMap = reactScopeListenerStore.get(reactScope);
+      if (eventTypeMap !== undefined) {
+        const type = ((event.type: any): DOMTopLevelEventType);
+        const listeners = eventTypeMap.get(type);
+        if (listeners !== undefined) {
+          const captureListeners = Array.from(listeners.captured);
+          const bubbleListeners = Array.from(listeners.bubbled);
+          const lastCurrentTarget = ((lastHostComponent: any): Element);
+
+          for (let i = 0; i < captureListeners.length; i++) {
+            const listener = captureListeners[i];
+            const {callback} = listener;
+            dispatchListeners.unshift(callback);
+            dispatchInstances.unshift(instance);
+            dispatchCurrentTargets.unshift(lastCurrentTarget);
+          }
+          for (let i = 0; i < bubbleListeners.length; i++) {
+            const listener = bubbleListeners[i];
+            const {callback} = listener;
+            dispatchListeners.push(callback);
+            dispatchInstances.push(instance);
+            dispatchCurrentTargets.push(lastCurrentTarget);
+          }
+        }
+      }
+    }
+    instance = instance.return;
+  }
+
+  // To prevent allocation to the event unless we actually
+  // have listeners we check the length of one of the arrays.
+  if (dispatchListeners.length > 0) {
+    event._dispatchListeners = dispatchListeners;
+    event._dispatchInstances = dispatchInstances;
+    event._dispatchCurrentTargets = dispatchCurrentTargets;
+  }
+}
+
+export function accumulateEventTargetListeners(
+  event: ReactSyntheticEvent,
+  currentTarget: EventTarget,
+): void {
+  const dispatchListeners = [];
+  const dispatchInstances: Array<Fiber | null> = [];
+  const dispatchCurrentTargets = [];
+
+  const eventTypeMap = eventTargetEventListenerStore.get(currentTarget);
+  if (eventTypeMap !== undefined) {
+    const type = ((event.type: any): DOMTopLevelEventType);
+    const listeners = eventTypeMap.get(type);
+    if (listeners !== undefined) {
+      const isCapturePhase = (event: any).eventPhase === 1;
+
+      if (isCapturePhase) {
+        const captureListeners = Array.from(listeners.captured);
+
+        for (let i = captureListeners.length - 1; i >= 0; i--) {
+          const listener = captureListeners[i];
+          const {callback} = listener;
+          dispatchListeners.push(callback);
+          // EventTarget listeners do not have instances, as there
+          // is no backing Fiber instance for them (window, document etc).
+          dispatchInstances.push(null);
+          dispatchCurrentTargets.push(currentTarget);
+        }
+      } else {
+        const bubbleListeners = Array.from(listeners.bubbled);
+
+        for (let i = 0; i < bubbleListeners.length; i++) {
+          const listener = bubbleListeners[i];
+          const {callback} = listener;
+          dispatchListeners.push(callback);
+          // EventTarget listeners do not have instances, as there
+          // is no backing Fiber instance for them (window, document etc).
+          dispatchInstances.push(null);
+          dispatchCurrentTargets.push(currentTarget);
+        }
+      }
+    }
+  }
+  // To prevent allocation to the event unless we actually
+  // have listeners we check the length of one of the arrays.
+  if (dispatchListeners.length > 0) {
+    event._dispatchListeners = dispatchListeners;
+    event._dispatchInstances = dispatchInstances;
+    event._dispatchCurrentTargets = dispatchCurrentTargets;
+  }
+}
+
+function getParent(inst: Fiber | null): Fiber | null {
+  if (inst === null) {
+    return null;
+  }
+  do {
+    inst = inst.return;
+    // TODO: If this is a HostRoot we might want to bail out.
+    // That is depending on if we want nested subtrees (layers) to bubble
+    // events to their parent. We could also go through parentNode on the
+    // host node but that wouldn't work for React Native and doesn't let us
+    // do the portal feature.
+  } while (inst && inst.tag !== HostComponent);
+  if (inst) {
+    return inst;
+  }
+  return null;
+}
+
+/**
+ * Return the lowest common ancestor of A and B, or null if they are in
+ * different trees.
+ */
+function getLowestCommonAncestor(instA: Fiber, instB: Fiber): Fiber | null {
+  let nodeA = instA;
+  let nodeB = instB;
+  let depthA = 0;
+  for (let tempA = nodeA; tempA; tempA = getParent(tempA)) {
+    depthA++;
+  }
+  let depthB = 0;
+  for (let tempB = nodeB; tempB; tempB = getParent(tempB)) {
+    depthB++;
+  }
+
+  // If A is deeper, crawl up.
+  while (depthA - depthB > 0) {
+    nodeA = getParent(nodeA);
+    depthA--;
+  }
+
+  // If B is deeper, crawl up.
+  while (depthB - depthA > 0) {
+    nodeB = getParent(nodeB);
+    depthB--;
+  }
+
+  // Walk in lockstep until we find a match.
+  let depth = depthA;
+  while (depth--) {
+    if (nodeA === nodeB || (nodeB !== null && nodeA === nodeB.alternate)) {
+      return nodeA;
+    }
+    nodeA = getParent(nodeA);
+    nodeB = getParent(nodeB);
+  }
+  return null;
+}
+
+function accumulateEnterLeaveListenersForEvent(
+  event: ReactSyntheticEvent,
+  target: Fiber,
+  common: Fiber | null,
+  capture: boolean,
+): void {
+  const registrationName = event.dispatchConfig.registrationName;
+  if (registrationName === undefined) {
+    return;
+  }
+  const dispatchListeners = [];
+  const dispatchInstances: Array<Fiber | null> = [];
+  const dispatchCurrentTargets = [];
+
+  let instance = target;
+  while (instance !== null) {
+    if (instance === common) {
+      break;
+    }
+    const {alternate, stateNode, tag} = instance;
+    if (alternate !== null && alternate === common) {
+      break;
+    }
+    if (tag === HostComponent && stateNode !== null) {
+      const currentTarget = stateNode;
+      if (capture) {
+        const captureListener = getListener(instance, registrationName);
+        if (captureListener != null) {
+          // Capture listeners/instances should go at the start, so we
+          // unshift them to the start of the array.
+          dispatchListeners.unshift(captureListener);
+          dispatchInstances.unshift(instance);
+          dispatchCurrentTargets.unshift(currentTarget);
+        }
+      } else {
+        const bubbleListener = getListener(instance, registrationName);
+        if (bubbleListener != null) {
+          // Bubble listeners/instances should go at the end, so we
+          // push them to the end of the array.
+          dispatchListeners.push(bubbleListener);
+          dispatchInstances.push(instance);
+          dispatchCurrentTargets.push(currentTarget);
+        }
+      }
+    }
+    instance = instance.return;
+  }
+  // To prevent allocation to the event unless we actually
+  // have listeners we check the length of one of the arrays.
+  if (dispatchListeners.length > 0) {
+    event._dispatchListeners = dispatchListeners;
+    event._dispatchInstances = dispatchInstances;
+    event._dispatchCurrentTargets = dispatchCurrentTargets;
+  }
+}
+
+export function accumulateEnterLeaveListeners(
+  leaveEvent: ReactSyntheticEvent,
+  enterEvent: ReactSyntheticEvent,
+  from: Fiber | null,
+  to: Fiber | null,
+): void {
+  const common = from && to ? getLowestCommonAncestor(from, to) : null;
+
+  if (from !== null) {
+    accumulateEnterLeaveListenersForEvent(leaveEvent, from, common, false);
+  }
+  if (to !== null) {
+    accumulateEnterLeaveListenersForEvent(enterEvent, to, common, true);
   }
 }
