+++ packages/react-dom/src/events/plugins/ModernBeforeInputEventPlugin.js
@@ -28,7 +28,7 @@ import {
 } from '../FallbackCompositionState';
 import SyntheticCompositionEvent from '../SyntheticCompositionEvent';
 import SyntheticInputEvent from '../SyntheticInputEvent';
-import accumulateTwoPhaseListeners from '../accumulateTwoPhaseListeners';
+import {accumulateTwoPhaseListeners} from '../DOMModernPluginEventSystem';
 
 const END_KEYCODES = [9, 13, 27, 32]; // Tab, Return, Esc, Space
 const START_KEYCODE = 229;
