+++ packages/react-dom/src/events/plugins/__tests__/ModernEnterLeaveEventPlugin-test.internal.js
@@ -11,6 +11,7 @@
 
 let React;
 let ReactDOM;
+let ReactFeatureFlags;
 
 describe('EnterLeaveEventPlugin', () => {
   let container;
@@ -18,6 +19,8 @@ describe('EnterLeaveEventPlugin', () => {
   beforeEach(() => {
     jest.resetModules();
 
+    ReactFeatureFlags = require('shared/ReactFeatureFlags');
+    ReactFeatureFlags.enableModernEventSystem = true;
     React = require('react');
     ReactDOM = require('react-dom');
