+++ packages/react-dom/src/events/plugins/__tests__/ModernSelectEventPlugin-test.internal.js
@@ -11,11 +11,14 @@
 
 let React;
 let ReactDOM;
+let ReactFeatureFlags;
 
 describe('SelectEventPlugin', () => {
   let container;
 
   beforeEach(() => {
+    ReactFeatureFlags = require('shared/ReactFeatureFlags');
+    ReactFeatureFlags.enableModernEventSystem = true;
     React = require('react');
     ReactDOM = require('react-dom');
